# api/docs/admin/admin_swagger.yaml
basePath: /api
definitions:
...
  admin.AuditConnLogIds:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
  admin.AuditFileLogIds:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  model.AddressBookCollection:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - name
    type: object
  model.AddressBookCollectionList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AddressBookCollection'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  model.AuditConn:
    properties:
      action:
        type: string
      close_time:
        type: integer
      conn_id:
        type: integer
      created_at:
        type: string
      from_name:
        type: string
      from_peer:
        type: string
      id:
        type: integer
      ip:
        type: string
      peer_id:
        type: string
      session_id:
        type: string
      type:
        type: integer
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  model.AuditConnList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AuditConn'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  model.AuditFile:
    properties:
      created_at:
        type: string
      from_name:
        type: string
      from_peer:
        type: string
      id:
        type: integer
      info:
        type: string
      ip:
        type: string
      is_file:
        type: boolean
      num:
        type: integer
      path:
        type: string
      peer_id:
        type: string
      type:
        type: integer
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  model.AuditFileList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AuditFile'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
  title: API
  version: "1.0"
paths:
...
  /admin/address_book_collection/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Address book collection form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/address_book_collection/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Address book collection deletion form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/address_book_collection/detail/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/address_book_collection/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number for the address book collection list
        in: query
        name: page
        type: integer
      - description: Number of collections per page
        in: query
        name: page_size
        type: integer
      - description: Filter for collections owned by the current user
        in: query
        name: is_my
        type: integer
      - description: id
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollectionList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/address_book_collection/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: Address book collection update form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
...
  /admin/audit_conn/batchDelete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Audit connection log ID batch deletion payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.AuditConnLogIds'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/audit_conn/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Audit connection deletion form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AuditConn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/audit_conn/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number for audit connection records
        in: query
        name: page
        type: integer
      - description: Number of audit connection records per page
        in: query
        name: page_size
        type: integer
      - description: Peer ID filter for audit connections
        in: query
        name: peer_id
        type: integer
      - description: Source peer ID filter for audit connections
        in: query
        name: from_peer
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AuditConnList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/audit_file/batchDelete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Audit file log ID batch deletion payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.AuditFileLogIds'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/audit_file/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Audit file deletion form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AuditFile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/audit_file/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number for audit file records
        in: query
        name: page
        type: integer
      - description: Number of audit file records per page
        in: query
        name: page_size
        type: integer
      - description: Peer ID filter for audit file records
        in: query
        name: peer_id
        type: integer
      - description: Source peer ID filter for audit file records
        in: query
        name: from_peer
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AuditFileList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"